Changing demo format to ASWebAuthenticationSession. Updating all apis and greatly reducing app bloat
