--- .github/workflows/build-DocC.yml
@@ -1,50 +1,0 @@
-# Simple workflow for deploying static content to GitHub Pages
-name: Deploy DocC content to Pages
-
-on:
-  # Runs on pushes targeting the default branch
-  push:
-    branches: ["main"]
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
-concurrency:
-  group: "pages"
-  cancel-in-progress: true
-
-jobs:
-  # Single deploy job since we're just deploying
-  deploy:
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: macos-13
-    steps:
-      - name: Checkout 🛎️
-        uses: actions/checkout@v3
-      - name: Build DocC
-        run: |
-          xcodebuild docbuild -scheme iGamingKit \
-            -derivedDataPath /tmp/docbuild \
-            -destination 'generic/platform=iOS';
-          $(xcrun --find docc) process-archive \
-            transform-for-static-hosting /tmp/docbuild/Build/Products/Debug-iphoneos/iGamingKit.doccarchive \
-            --hosting-base-path iGaming-iOS \
-            --output-path docs;
-          echo "<script>window.location.href += \"/documentation/igamingkit\"</script>" > docs/index.html;
-      - name: Upload artifact
-        uses: actions/upload-pages-artifact@v1
-        with:
-          # Upload only docs directory
-          path: 'docs'
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v1
