--- Package.resolved
@@ -1,14 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "plaid-link-ios",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/plaid/plaid-link-ios",
-      "state" : {
-        "revision" : "c675e4b4cd8c72d2c1f2cffdef6c81b0574d365c",
-        "version" : "4.7.0"
-      }
-    }
-  ],
-  "version" : 2
-}
