--- documentation/igamingkit/pavilionwebviewconfiguration/pavilionwebviewdidcomplete/index.html
@@ -1,1 +1,0 @@
-<!DOCTYPE html><html lang="en-US"><head><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="viewport" content="width=device-width,initial-scale=1,viewport-fit=cover"><link rel="icon" href="/https://github.com/PavilionPay/iGaming-iOS/favicon.ico"><link rel="mask-icon" href="/https://github.com/PavilionPay/iGaming-iOS/favicon.svg" color="#333333"><title>Documentation</title><script>var baseUrl = "/https://github.com/PavilionPay/iGaming-iOS/"</script><link href="/https://github.com/PavilionPay/iGaming-iOS/css/chunk-c0335d80.10a2f091.css" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/css/documentation-topic.1d1eec04.css" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/css/documentation-topic~topic.b6287bcf.css" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/css/documentation-topic~topic~tutorials-overview.d6f5411c.css" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/css/topic.d8c126f3.css" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/css/tutorials-overview.c249c765.css" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/chunk-2d0d3105.cd72cc8e.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/chunk-c0335d80.76a68cc5.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/documentation-topic.57e91f8a.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/documentation-topic~topic.1679ec90.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/documentation-topic~topic~tutorials-overview.90c61522.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-bash.1b52852f.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-c.d1db3f17.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-cpp.eaddddbe.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-css.75eab1fe.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-custom-markdown.7cffc4b3.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-custom-swift.5cda5c20.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-diff.62d66733.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-http.163e45b6.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-java.8326d9d8.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-javascript.acb8a8eb.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-json.471128d2.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-llvm.6100b125.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-markdown.90077643.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-objectivec.bcdf5156.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-perl.757d7b6f.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-php.cc8d6c27.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-python.c214ed92.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-ruby.f889d392.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-scss.62ee18da.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-shell.dd7f411f.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-swift.84f3e88c.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/highlight-js-xml.9c3688c7.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/topic.8cd0c0c4.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/tutorials-overview.2a32cd6f.js" rel="prefetch"><link href="/https://github.com/PavilionPay/iGaming-iOS/css/index.038e887c.css" rel="preload" as="style"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/chunk-vendors.ba2dd0cb.js" rel="preload" as="script"><link href="/https://github.com/PavilionPay/iGaming-iOS/js/index.e8a5d294.js" rel="preload" as="script"><link href="/https://github.com/PavilionPay/iGaming-iOS/css/index.038e887c.css" rel="stylesheet"></head><body data-color-scheme="auto"><noscript><style>.noscript{font-family:"SF Pro Display","SF Pro Icons","Helvetica Neue",Helvetica,Arial,sans-serif;margin:92px auto 140px auto;text-align:center;width:980px}.noscript-title{color:#111;font-size:48px;font-weight:600;letter-spacing:-.003em;line-height:1.08365;margin:0 auto 54px auto;width:502px}@media only screen and (max-width:1068px){.noscript{margin:90px auto 120px auto;width:692px}.noscript-title{font-size:40px;letter-spacing:0;line-height:1.1;margin:0 auto 45px auto;width:420px}}@media only screen and (max-width:735px){.noscript{margin:45px auto 60px auto;width:87.5%}.noscript-title{font-size:32px;letter-spacing:.004em;line-height:1.125;margin:0 auto 35px auto;max-width:330px;width:auto}}#loading-placeholder{display:none}</style><div class="noscript"><h1 class="noscript-title">This page requires JavaScript.</h1><p>Please turn on JavaScript in your browser and refresh the page to view its content.</p></div></noscript><div id="app"></div><script src="/https://github.com/PavilionPay/iGaming-iOS/js/chunk-vendors.ba2dd0cb.js"></script><script src="/https://github.com/PavilionPay/iGaming-iOS/js/index.e8a5d294.js"></script></body></html>
