+++ iGamingDemo.xcodeproj/xcshareddata/xcschemes/iGamingDemo.xcscheme
