+++ iGamingDemo/Assets.xcassets/Contents.json
