+++ iGamingDemo/BuildEnvironmentVariables.plist
@@ -2,15 +2,11 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>RedirectUri</key>
-	<string></string>
-	<key>BaseUri</key>
-	<string></string>
-	<key>Audience</key>
+	<key>Secret</key>
 	<string></string>
 	<key>Issuer</key>
 	<string></string>
-	<key>Secret</key>
+	<key>Environment</key>
 	<string></string>
 </dict>
 </plist>
