+++ iGamingDemo/Info.plist
