+++ iGamingDemo/SessionSetupViewController.xib
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="32700.99.1234" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22685"/>
+        <capability name="Named colors" minToolsVersion="9.0"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="SessionSetupViewController" customModule="iGamingDemo" customModuleProvider="target">
+            <connections>
+                <outlet property="activityIndicator" destination="LSO-90-Y1X" id="WoP-bq-CaL"/>
+                <outlet property="launchButton" destination="FO2-FW-5lD" id="K5J-9q-UoZ"/>
+                <outlet property="view" destination="khw-S5-hcF" id="1IB-JR-drY"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="khw-S5-hcF">
+            <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="iGaming Demo App" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="wpI-Gj-gqj">
+                    <rect key="frame" x="123.00000000000001" y="139" width="147.33333333333337" height="21"/>
+                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                    <nil key="textColor"/>
+                    <nil key="highlightedColor"/>
+                </label>
+                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="FO2-FW-5lD">
+                    <rect key="frame" x="104.33333333333333" y="421" width="184.33333333333337" height="35"/>
+                    <state key="normal" title="Button"/>
+                    <buttonConfiguration key="configuration" style="plain" title="Launch iGaming SDK">
+                        <backgroundConfiguration key="background">
+                            <color key="backgroundColor" name="AccentColor"/>
+                        </backgroundConfiguration>
+                        <color key="baseForegroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    </buttonConfiguration>
+                    <connections>
+                        <action selector="launchVipSDK" destination="-1" eventType="touchUpInside" id="OfE-Ma-Ivo"/>
+                    </connections>
+                </button>
+                <activityIndicatorView hidden="YES" opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" hidesWhenStopped="YES" style="large" translatesAutoresizingMaskIntoConstraints="NO" id="LSO-90-Y1X">
+                    <rect key="frame" x="178" y="506" width="37" height="37"/>
+                </activityIndicatorView>
+            </subviews>
+            <viewLayoutGuide key="safeArea" id="Bdz-Dn-tgM"/>
+            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <constraints>
+                <constraint firstItem="wpI-Gj-gqj" firstAttribute="centerX" secondItem="Bdz-Dn-tgM" secondAttribute="centerX" id="EiM-8r-309"/>
+                <constraint firstItem="FO2-FW-5lD" firstAttribute="centerX" secondItem="Bdz-Dn-tgM" secondAttribute="centerX" id="Hz8-r5-YZs"/>
+                <constraint firstItem="LSO-90-Y1X" firstAttribute="top" secondItem="FO2-FW-5lD" secondAttribute="bottom" constant="50" id="Hza-6f-azg"/>
+                <constraint firstItem="FO2-FW-5lD" firstAttribute="centerY" secondItem="Bdz-Dn-tgM" secondAttribute="centerY" id="aeP-F2-dqz"/>
+                <constraint firstItem="wpI-Gj-gqj" firstAttribute="top" secondItem="Bdz-Dn-tgM" secondAttribute="top" constant="80" id="srv-3y-5aB"/>
+                <constraint firstItem="LSO-90-Y1X" firstAttribute="centerX" secondItem="FO2-FW-5lD" secondAttribute="centerX" id="uuz-No-yah"/>
+            </constraints>
+            <point key="canvasLocation" x="-184" y="-627"/>
+        </view>
+    </objects>
+    <resources>
+        <namedColor name="AccentColor">
+            <color red="0.0" green="0.46000000000000002" blue="0.89000000000000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
