+++ iGamingDemo/iGamingDemo.entitlements
