--- iGamingDemo/iGamingDemo.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,15 +1,0 @@
-{
-  "originHash" : "c0341d434e7350f2afe06ffcbd403770596937cd4a122e7901eb4264c8adf474",
-  "pins" : [
-    {
-      "identity" : "plaid-link-ios",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/plaid/plaid-link-ios",
-      "state" : {
-        "revision" : "296a3c21b1b8561ee3a9f859a3c8194043c843d0",
-        "version" : "5.6.0"
-      }
-    }
-  ],
-  "version" : 3
-}
